refactor: custom hook으로 관심사 분리
