+++ src/App.tsx
@@ -1,92 +1,47 @@
-import { useCallback, useState } from "react";
-import { generateItems } from "./utils";
-import type { User } from "./types/user";
-import type { Notification } from "./types/notification";
 import { Header } from "./components/Header";
-import { ItemList, type Item } from "./components/ItemList";
+import { ItemList } from "./components/ItemList";
 import { ComplexForm } from "./components/ComplexForm";
 import { NotificationSystem } from "./components/NotificationSystem";
+import { useItems, useTheme, useAuth, useNotifications } from "./hooks";
 
-// 메인 App 컴포넌트
 const App: React.FC = () => {
-  const [theme, setTheme] = useState("light");
-  const [items, setItems] = useState<Item[]>(() => generateItems(1000));
-  const [user, setUser] = useState<User | null>(null);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
+	const { theme, toggleTheme } = useTheme();
+	const { items, addItems } = useItems();
+	const { notifications, addNotification, removeNotification } =
+		useNotifications();
+	const { user, login, logout } = useAuth(addNotification);
 
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
-  }, []);
-
-  const addNotification = useCallback(
-    (message: string, type: Notification["type"]) => {
-      const newNotification: Notification = {
-        id: Date.now(),
-        message,
-        type,
-      };
-      setNotifications((prev) => [...prev, newNotification]);
-    },
-    [],
-  );
-
-  const login = useCallback(
-    (email: string) => {
-      setUser({ id: 1, name: "홍길동", email });
-      addNotification("성공적으로 로그인되었습니다", "success");
-    },
-    [addNotification],
-  );
-
-  const logout = useCallback(() => {
-    setUser(null);
-    addNotification("로그아웃되었습니다", "info");
-  }, [addNotification]);
-
-  const removeNotification = useCallback((id: number) => {
-    setNotifications((prev) =>
-      prev.filter((notification) => notification.id !== id),
-    );
-  }, []);
-
-  return (
-    <div
-      className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-    >
-      <Header
-        theme={theme}
-        toggleTheme={toggleTheme}
-        user={user}
-        login={login}
-        logout={logout}
-      />
-      <div className="container mx-auto px-4 py-8">
-        <div className="flex flex-col md:flex-row">
-          <div className="w-full md:w-1/2 md:pr-4">
-            <ItemList items={items} onAddItemsClick={addItems} theme={theme} />
-          </div>
-          <div className="w-full md:w-1/2 md:pl-4">
-            <ComplexForm
-              user={user}
-              notifications={notifications}
-              addNotification={addNotification}
-            />
-          </div>
-        </div>
-      </div>
-      <NotificationSystem
-        notifications={notifications}
-        removeNotification={removeNotification}
-      />
-    </div>
-  );
+	return (
+		<div
+			className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+		>
+			<Header
+				theme={theme}
+				toggleTheme={toggleTheme}
+				user={user}
+				login={login}
+				logout={logout}
+			/>
+			<div className="container mx-auto px-4 py-8">
+				<div className="flex flex-col md:flex-row">
+					<div className="w-full md:w-1/2 md:pr-4">
+						<ItemList items={items} onAddItemsClick={addItems} theme={theme} />
+					</div>
+					<div className="w-full md:w-1/2 md:pl-4">
+						<ComplexForm
+							user={user}
+							notifications={notifications}
+							addNotification={addNotification}
+						/>
+					</div>
+				</div>
+			</div>
+			<NotificationSystem
+				notifications={notifications}
+				removeNotification={removeNotification}
+			/>
+		</div>
+	);
 };
 
 export default App;
