+++ src/hooks/useNotifications.ts
@@ -1,0 +1,27 @@
+import { useCallback, useState } from "react";
+import type { Notification } from "../types/notification";
+
+// 알림 관련 로직
+export function useNotifications() {
+	const [notifications, setNotifications] = useState<Notification[]>([]);
+
+	const addNotification = useCallback(
+		(message: string, type: Notification["type"]) => {
+			const newNotification: Notification = {
+				id: Date.now(),
+				message,
+				type,
+			};
+			setNotifications((prev) => [...prev, newNotification]);
+		},
+		[],
+	);
+
+	const removeNotification = useCallback((id: number) => {
+		setNotifications((prev) =>
+			prev.filter((notification) => notification.id !== id),
+		);
+	}, []);
+
+	return { notifications, addNotification, removeNotification };
+}
